Bug 696364: Restrict size of alphabits buffer
Thanks to Michael for doing the investigation, and leaving a
really good write up of the problem on the bug.

The basic problem is that we have a transparency group that
contains a softmask. The softmask is larger than the transparency
group, so it is restricted to the size of the group and written
into the clist as occupying just a small set of bands.

The content for the softmask is larger, however, so when used with
GraphicsAlphaBits, we create various alphabits devices, draw
the contents, and then send the results with copy_alpha.

These copy_alpha calls are NOT restricted to being within the
reduced region for the softmask. This therefore upsets the clist
reading.

The fix, as suggested by Michael in his bug report is to limit
the size of the softmask contents (by limiting the size of the
alphabits buffers).

Ideally the alphabits device would know what region to limit
itself to just by looking at the graphics state it's passed in.
Unfortunately, that doesn't work as the current softmask/groups
aren't reflected in the graphic state. This feels wrong to me,
but that's the way it is.

Instead, we have to resort to asking the device to limit our
bbox according to the current state. We achieve this by adding
a new dev_spec_op, and adding the required plumbing to the pdf14
device (to pass it on to the target) and the clist device (to
actually do the restriction).