Bug 697372: Fix crash due to pattern clist being freed
When rendering a pattern from a clist into the pdf14 compositor, the compositor
takes a reference to the pattern clist device. This is fine pre-page-clist, or
with no page clist involved (when the memory is managed by the garbage
collector), but post-page-clist the life span of the pattern clist device is
dicated by the clist replay, *not* by references to it (as in the gc'ed case).
This can result in a dangling pointer, which later causes a crash in the
garbage collector when it cleans up after the page-clist has completed.

So, when pattern fill has been completed, remove the reference to the pattern
clist device.