Fix clump handling.
Ray spotted that running:

  debugbin/gs -r300 -Z@\$\?: -dJOBSERVER

is enough to send us into an infinite loop.

Tracking this down, it appears that the reason is to do with
the gs_ref_memory_t's use of 'pcc' and 'cc'.

It copies the 'current' chunk out of the splay tree, and then
puts it back later, without realising that the splay tree
may have been altered.

The simplest fix appears to be just to not do this copying.