fix bug 697621: color_usage.or bits were wrong with transparency
The psdcmykog uses the color_usage.or bits to decide which components need to
be processed, skipping those that are not needed. When the pdf14 device changes
the color_info the color that will actually go on the page cannot be known
without transforming the color to the device color which is too expensive.

Also fix the polarity tests in gx_color_index2usage (it was being ignored)