Bug 697545: Extend "separable_and_linear" processing.
First, avoid needless check_separable_and_linear tests.

When we make a memory store for a target device, we use
the encoding/decoding color functions from that target
device. It seems reasonable to copy the "separable_and_linear"
information from that device too, to avoid having to needlessly
retest using the same encoding/decoding functions to get
the same answer.

Next, extend gx_color_enc_sep_lin_t to represent not just
whether we know details about the device colors separability
and linearity (and what those details are if known), but also
to encode whether those are standard encodings or not. The
standard encoding is defined to be "that which is compatible
with the pdf14 compositor".

Finally, we add a function, check_device_separable_encoding,
that 'promotes' the record for a given space from being a
known separable_and_linear one, to one that knows it's
separable_and_linear, but also knows whether it uses the
standard encoding.

Call this in clist_close_writer_and_init_reader to ensure that
the buffer device is created knowing whether it's a standard
encoding or not. If called here, it only ever needs to run once.

Call it again in pdf14_ok_to_optimize (just in case we ever slip
through - this takes no time if it has been run already). This
enables us to make a correct choice, and to avoid skipping
transparency bands if we are not using the standard (pdf14
compositor compatible) encoding.