txtwrite - fix Unicode conversion of glyph names 'unixxxx'
More fallout from the ToUnicode revamp required for pdfwrite, commit
9dba57f0f9a53c130ec2771c0ed1d7bd6bbef6ab

The code to handle converting glyphs whose name conforms to the
uniXXXX form, when there is no better ToUnicode information, was still
returning the number of bytes, not shorts. In addition, the original
change was packing the 2 bytes as two shorts instead of a singe short.

This resulted in incorrect output. Fixed here by packing the 4 nibbles
into a single short and returning a count of 1 short.