Fix bug 695280 page mode and clist mode differ
The clist mode "uses_color" was ignoring CombineWithColor if the rop
did not use the texture (rop3_uses_T false). This affects the ps3cet
09-40.PS and pcl5ccet/23-11.BIN files.

Note that this reults in these cases not using high-level images
in the clist since currently the clist cannot accomodate the two
different colorspaces (painting an Indexed colorspace image through
a pattern). Since the number of changed file is minimal, I assume
that the performance impact does not affect many files.