pdfwrite: properly handle text outside the clip
Bug 697442:

To work around an Acrobat limitation (related to glyphs with very large
coordinates), we don't emit glyph whose bounding box is outside the current
clipping region.

However, the file in question here uses a combination of show and glyphshow to
display text, and it also measures (in the manner of stringwidth) to position
the text. *But* because it relies on glyphshow, it cannot use stringwidth. It
moves the current point outside the page, draws the text (strings and glyph
names), and records how far the current point has moved.

The previous code would skip out early to avoid emitting the clipped out glyphs,
whilst still apparently updating the current point. *But* the values from which
the current point were updated were not correctly set due to skipping out
early.

We now do everything *except* emit the glyphs, thus the various values are
correct, and the currentpoint is updated properly.