PDF interpreter - add B operators to spcial 'W' handling
Previously we had altered the handling of stroke, fill and eofill when
in a 'W' context, to detect changes in the graphics state between the
'W' or 'W*' operation and the terminating path operator.

Unfortunately when this was done the B, b, B* and b* operators were
missed, which meant these always behaved as if an invalid graphics
state change was executed, ignored the graphics state change, and
ended up rednering incorrectly.

This commit adds those operators to the special handling of the W and
W* operators.