Set tags correctly for glyphs when sent as masked images.
In gx_image_cached_char we can end up sending the glyph as
an imagemask. When we do so, the tag is incorrectly reset
to image, rather than text.

Here we correct that by adding an 'image_is_text' flag to
gs_image_init and gs_image_begin_typed.

This can be seen by the fact that:

  gs -sDEVICE=bitrgbtags -o out.bit -r600 -dMaxBitmap=1000
    tests_private/customer_tests/jw202dwo.pcl

and

  gs -sDEVICE=bitrgbtags -o out.bit -r600 -dMaxBitmap=400000000
    tests_private/customer_tests/jw202dwo.pcl

give different tag values for the text.