Move invalid_file_stream into gc memory.
At initialisation we create an special stream (in the gs context) to use as an
invalid stream whenever we need one.

It was originally being allocated in non-gc memory. Since the context is in
gc memory, the invalid_file_stream can leak. Rather than add explicit freeing
of the invalid_file_stream, and risk a life span conflict between a gc object
and a non-gc object, it makes it neater, and more sense, to keep the
invalid_file_stream in gc memory. It *is* stable memory (not subject to
save/restore) and allocated as an immovable object (not subject to relocating
by the garbager).

Fixes a problem spotted by Shailesh Mistry while working on memory squeezing
event 170.