Bug 697681: fix shared lib build with openjpeg >= 2.1.1
In 2.1.1 openjpeg extended the use of the OPJ_STATIC to use gcc attributes to
hide certain functions, as well as the previous use of preventing the
inclusion of the Windows DllMain function.

As we (in Ghostscript) need the functions the OPJ_STATIC on Linux/Unix hides,
only set the flag for Windows builds.