pdfwrite - don't try to allocate a 0 length buffer for metadata
Bug #697684 "Regression - Ghostscript 9.21 fails when pdfmark contains certain Unicode characters"

Under some conditions, specifying a 0 length buffer size when calling
gs_alloc_bytes() can return a NULL pointer. Since we use a NULL return
to indicate a failure to allocate a buffer, this results in a VM error.

In this commit we check the data length and don't try to allocate or
use a 0 length buffer.

We'll tackle the allocation problem separately, it would be better if
gs_alloc_bytes didn't return a NULL pointer.