Fix clist file to invalidate the read cache upon writing (see bug 697689).
This fixes the unknownerror reported by the bug, but does not totally fix
the use of # false .outputpage with the clist since the data after the
first cmd_opv_end_page is not rendered to the page (stops the clist reader).
The bug is left open for that reason and has been marked bountiable.