Fix a problem with tags not being set in dev_color
When the tag is changed, we need to unset the dev_color so that the
device's encode_color gets called to apply the correct tag. Seen with
the bitrgbtags device and the file:  tests/pdf/Bug691997.pdf

Also, fix a problem with the hit_detect device (infill PS operator).
If the dev_color is unset in the fill or stroke operation, it would
attempt to remap the dev_color which would fail because the hit detect
device doesn't have a device profile. This sets the grapchis_type_tag
to GS_PATH_TAG so that fill and stroke don't unset (and subsequently
remap in gx_set_dev_color) the dev_color.