Bug 697678: Scan converter performance (out of range lines).
This significantly improves the timings.

Essentially, if we were rendering just a band of a page (say from
Y1 to Y2) and the line extended beyond that (say from Y0 to Y3,
where Y0 < Y1 or Y2 < Y3), we'd have run the bresenham for the
entire range rather than for just the section we needed to.

This was particularly problematic in the example file due to the
use of a clipping device, meaning that we were repeatedly
scan converting lines in 1 pixel high regions.