PDF interpreter - don't remove Filters from streams declared as 0 Length
No bug number for this, the problem arose from a Stack Overflow
question here http://stackoverflow.com/questions/43149372/repairing-pdfs-with-damaged-xref-table/43154410?noredirect=1#comment73447166_43154410

The PDF file is multiply broken, it has invalid xref entries and each
page is an image XObject whose dictionary has an indirect object for
the /Length entry. In every case the Length is defined as 0.

For some obscure reason, sometime in the past someone added code to
remove the /Filter entry from the dictionary when a stream has a
declared Length of 0. I have no idea why, if the stream is compressed
or encoded then removing the Filter can have no useful effect. But
if the Filter is left in place we can still recover from an invalid
Length declaration, provided we properly decode the stream, which
obviously requires us to install the correct Filter(s).

So here we simply remove the code which undefines the Filter in the
stream dictionary.

cluster runs show no differences.