Fix error in color updating
Ray found an issue that would appear to cause problems when we have
transparency and spot colors for the separations devices.
The mask and shift values in color info were not getting set properly
for the spot colors in this case.  It is not a real issue
due to the way that the separation device colors are
stored in the devn hl color (planar) but is good to fix
to avoid confusion.