Bug 697722: Skip glyph cache with vertical substitution
The way PXL implements vertical glyph substitution means it defeats our glyph
cache logic - the cache doesn't have the information to differentiate between
a vertical substitute and a regular glyph.

As this is a very rarely used feature, the simplest solution is to disable use
of the cache when vertical substitutes are in force - it will neither lookup
existing, nor add new glyph instances to the cache.

If we start seeing more jobs using vertical glyph substitution, and this proves
to be a performance bottleneck, we can revisit this and add the required data
for the cache to handle these (and possibly other) cases.