(e)ps2write - use correct encoding for subset TrueType fonts
Bug #697733 "Character section is not handled with eps2write"

We carry around a 'base encoding' with fonts, which is generally the
encoding of the original font, potentially 'refined' to a more useful
encoding if its originally one we can't use.

For TrueType fonts, we would either Standard or ISOLatin1, and in fact
almost always StandardEncoding.

The problem is that if we are subsetting the font, then we *don't* embed
the original font's CMAP subtable, we generate a new pair, one of Windows
Unicode (3,1) and one of Mac Roman (1,0). Neither is suitable for use
with StandrdEncoding and in fact the opdfread code will always use the
MacRoman Encoding.

This can lead to a problem if the same glyph name is encoded in one or
other Encodings, and is either missing or in a in a different position
in the other.

When we encode the string we will use StandardEncoding, and the opdfread
code will read the character codes using MacRoman Encoding, so the
Charstrings dictionary will be incorrectly constructed and glyphs will
either become notdef or potentially use the wrong glyph.

To fix it; if we have a TrueType font, and are subsetting it, select
a Base Encoding of MacRoman. Update the refine_encoding procedure to
permit a TrueType fotn to have MacRoman encoding.

This shows a minor progression in one file in the test suite;
Bug691848.ps, surprisingly with Kanji text.....