PCL - remove the 'bias' from glyph indices before use, if present
Bug #697744 "Accents not handled correctly"

In plfont.c, pl_glyph_name, we use the 'glyph' (which is of type
gs_glyph) as an index to retrieve a glyph name from a TrueType font.

However, as per the documentation in gsccode.h, if the glyph is truly
an index (as opposed to a name) then it is biased by GS_MIN_GLYPH_INDEX
which means that attempting to use it directly as an index into the font
fails, and returns no name.

This causes pdfwrite multiple problems; firstly it causes the original
glyph to be rendered as a bitmap. Then, because the glyph is a composite
glyph (its made up of two or more other glyphs) subsequent uses of the
same glyph cause it to become confused, and use the initail component
glyph.

This can be easily resolved simply by testing to see if the glyph is
biased by GS_MIN_GLYPH_INDEX and, if it is, removing the bias.

In addition to fixing the reported problem a number of test suite files
show progressions (accents now present) and small but useful quality
improvements (glyphs no longer rendered as bitmaps) as well as being
able to properly extract text from either the PCL or the output PDF file.