subclassing devices - process the graphics_type_tag into all chained devices
The tag devices encode the type of operation into the device color, if
the devices graphics_type_tag has the GS_DEVICE_ENCODES_TAGS bit set.

The specific tag is maintained by altering the graphics_type_tag in the
device structure using a device method (set_graphics_type_tag). However,
when checking to see if the tag needs to change (for a different marking
operation type), the code does not, as might be expected, use a
symmetric device method such as get_graphics_type_tag. It was felt that
the performance impact would be unacceptable. Instead the
graphics_type_tag of the *current* device in the graphics state is
inspected directly from the structure.

This means that every device chaining method (and there are several,
see ghostpdl/doc/subclass.htm) needs to ensure that every device in
the chain sets its graphics_type_tag in the device structure, *and*
passes that on to any underlying devices.

Apparently the 'forwarding' devices already do this, I presume the clist
and transparency compositors do as well, I haven't checked. The obselete
'device filter chain' does not, the subclassing devices did not either.

Its also important (for pattern colour spaces) that we track the tag
even when we are not actually rendering.

So here we remove the code from the First/Last page device which skipped
tags on elided pages. This means the device can instead use the standard
subclassing default method.

In the standard subclassing default method we now update the tag on the
current device in the chain before passing it on to the next device.
This means that all devices in the chain now track the tag. An important
note for any subclassing devices which implement a set_graphics_type_tag
method is that it must set the tag in its device structure. So we document
that as well.