Tidy up, remove debugging prints
A cluster push of this code exhibits diffs, mostly in the 'tags' devices
which I've chosen to ignore (its hard to see how colour changes should
affect the tags), not least because this device is so flaky.

The remaining diffs fall into two camps; firstly progressions, a small
number of the files are now correct (or better). Secondly a couple of
files which are invalid and which we now fail on in slightly different
ways to before, which I don't propose to worry about. One is now a
better match for Acrobat, the other is worse.