PDF interpreter - detect missing 'endobj' on final object in file
Bug #697761 " Broken PDF (missing endobj on final object) causes error"

The sample  file has the final two objects (Info and Catalog dictionaries)
written with missing 'endobj'  Our normal processing catches that in the
body of the file, but if its the final object, then we carried on and
tried to process the entire trailer and xref.

Add 'xref' to the entries in resolveopdict and treat it as a missing
endobj. While we're here, add an error message for the case of detecting
an 'obj' when resolving an object. We already handle that, but we
weren't producing an error for it.