Bug 697762: Fix scan converter SEGVs.
The SEGVs are caused by the scan converter allowing less room
in the table than is actually required. This happens due to
the mark_curve subdivision overflowing.

For example, if we have a curve such as:

 0 0x80000000 moveto
 0 0x80000000 0 0x80000000 0 0x80000000 curveto

Then the current subdivision logic does:

 fixed ay = (sy + c1y)>>1

For sy = c1y = 0x80000000 this gives ay = 0, when clearly
the average should be 0x80000000.

An alternative formulation of the splitting code:

 fixed ay = (sy>>1) + (c1y>>1)

gives the required result, at the expense of dropping accuracy at
the low end.

This commit therefore has 2 formulations of the splitting code, one
for 'big' values (values that might overflow), and ones for normal
values (values that won't).

We select between the two pieces of code by comparing the top 2
bits of each value; if they differ, then we use the big variant.

Also, when calculating the size of the table required, we were
incorrectly closing the curve more often than we needed to. This
had no ill effect other than inflating the amount of memory we
needed.