Initialise path bounding boxes.
Bug #697770 "Inconsistent behaviour"

The problem is that gx_path_init_bbox_accumulator() sets bbox_accurate
which is later used to signal that the attached bbox is valid and
accurate.

But we don't actually initialise the bbox member at all, in the bug in
question the glyph has no contours, so creates no path, leaving the
bbox uninitialised. But because we have set bbox_accurate we just
copy the values when determining the font bounding box, leading to
insane values.

Initialising the bbox gives more reasonable results.