PS interpreter - restore correct error behaviour for invalid /Decode arrays
PDF interpreter - trim excessively long Decode arrays to the correct length

Bug #697787 " regression, several EPS figures which work fine with gs <= 9.07 throw Error with gs > 9.08"

commit 89c224b029357002c48660dae98de2d1e560d4e0 was applied to fix
bug #694276, and the fix was, I'm afraid, incorrect.

The bug in question was a PDF file where the /Decode array was too long
(by 1 entry) for the actual colour space. This caused the PostScript
interpreter to (correctly) throw an error. Ray fixed this by altering
the PostScript interpreter so that it did not throw an error when the
Decode array was too long, but instead used the requisite number of
elements.

This is not correct, the PostScirpt interpreter should throw an error.

The file attached to this report is an EPS file which relies (horribly)
on getting an error when the Deocde array is too long for an image, and
uses that to call colorimage instead.

By not throwing an error we treat the image as being in DeviceGray and,
since the data is actually RGB, only read 1/3 of the image data. We then
attempt to carry on from that point and throw real errors due to reading
from the middle of a data string.

This commit reverts (significant) the changes in commit
89c224b029357002c48660dae98de2d1e560d4e0. It also addresses the 'bug'
in report #694726 by shifting the checking to the PDF interpreter and
having that correctly set the decode array.