PDF interpreter - Apply /Widths to illegal (empty) type 3 CharProcs
Bug #697805 "Incorrect spacing between text"

The file (created by the Skia graphics library) embeds all the fonts as
type 3, the CharProc for the space glyph is embedded as an empty stream:

17 0 obj
<<
  /Length 0
>>
stream
endstream
endobj

The specification does not say what should happen if the /Width entry
is not the same as the actual width in the glyph description, but we
choose to use the /Width, because from prior experience we know that's
what Acrobat does.

However, this glyph is illegal, because it contains neither a d0 nor a d1
operator. We apply the Widths override when executing d0 or d1, so if
the (required) operator is not present in the glyph description we
cannot override it.

In this commit we detect a 0 length stream and treat it the same as a
missing stream; we execute a dummy d1 which enables us to do the
Widths override.