Change TTF cmap procedure name and tidy up
The procedure for parsing the TTF cmap table was called "cmaparray", but it
has returned a dictionary rather than an array for a *long* time. So change
the procedure name to "cmapdict".

In the format 4 cmap parsing, left over from when we did return an array, we
looped over the codes, counting as we go, so we could allocate an array large
enough to hold all the mappings, then looped over them again to fill in the
array.

As we now use a dictionary, there is no need to pre-count, so the first
loop was redundant - remove it.