Bug 697769: cmap tab for symbolic fonts, and format 4 cmap parsing
This involved two problems:
Firstly, the logic of whether to treat a font flagged as symbolic *really*
as symbolic was missing a condition. It appears that as well as checking for
a symbolic (3.0) cmap table, we also need to check for a UCS-2 (3.1) cmap
table.

As noted before, this area is poorly documented in the spec, so it is possible
further issues may arise.

Secondly, the fonts in the test file are, in fact, invalid. Each using a format
4 segmented cmap table with only one, non-terminal, segment. According to the
TTF spec, a format 4 table should have a terminating segment whose start and
end codes are 0xffff. These fonts do not have such a segment. As it happens,
we ignore the terminal segment anyway, so by simply tweaking the loop
condition, we can parse the one segment, and retrieve a usable cmap.