Correct function /Domain array checks
Bug #697858 "Function Domain validation incorrect"

The PDF and PostScript specifications state that the entries in a Domain
array for a function must be such that Domain2i <= Domain2i+1

However the code in fn_build_sub_function() inverted the sense by
checking if Domain2i >= Domain2i+1 and throwing an error if so. This is
of course incorrect for the case where Domain2i == Domain2i+1, this
should not be an error, but the code treated it as an error.