Bug 697861 Avoid push of text transparency group when no text is drawn
The text can be handed to the pdf14 device text_begin operation with a graphic
state that has a null device.  This occurs when the text state was
set to do no stroke or fill, only movements.   In such a case we
don't want to do the text group push until we actually have some
text to draw.