Bug 697813: Fix integer overflow leading to buffer overflow.
When calculating the amount of space required for a buffer, we
were using an int. This would overflow, meaning we'd allocate
less than we needed, and then segmentation fault when we overran
the end.

Fix is to count the space required in an int64_t, and check that
that actually fits into a uint (the type passed to the allocation
function) before allocating.

Side note: Surely gs should be using size_t's rather than uints
for its allocation calls?

Also, our error cleanup was freeing the wrong thing. Fix that too.