Bug 697847: Fix SEGV seen in scan converter.
Previously the scan converter had had problems with extremely
large/small values over/underflowing during bezier subdivision.

We'd fixed this by calling into a "_big" variant of the
subdivision routine that sacrificed accuracy in the bottom bit.

Unfortunately, this can cause 'jitter' in the low bits of the
subdivided curves, which can cause scanlines to be crossed
repeatedly. Once a single curve can cross a given scanline
more times than expected, our buffer allocation is too small
and we get crashes such as the one seen here.

The fix is to move to using 64bit accuracy in the _big variant
and thus not introduce the jitter.