Bug 696422: bounds check number of points in TTF glyph
There was already a bounds check for the number of points in a TTF glyph
outline contour, to avoid buffer overflow, but it failed to check that the value
was positive. In this case, the value is negative, which caused loops to fail
to exit. This is sufficiently broken, we can only give up and return an error