Bug 696417: fully initialise TTF bytecode context.
Although the font and CVT programs aren't supposed to use certain opcodes,
specifically, in this case, the opcodes that reference the contents of the
zone pointers. But this font contains programs that do. Since other interpreters
handle this, just ensure we initialize the context fully (so the zone pointers
are valid) before executing the context.