ps2epsi - reapply fix
Bug #697890 "Bug 689832 still unresolved in master"

The fix for this was applied (contrary to the report in 697890) but it
was later removed as part of the fix for bug #688500.

I presume this was accidental, since 688500 predates 689832, but the
fix was applied later. I don' know how this happened and after 8 years
I don't propose to worry about it.

Just reapply the fix for 689832