pdfwrite - make sure an EP pdfmark is not an orphan
Bug #696375 "Seg faults found by fuzzing in cos_object_memory (gdevpdfo.c:189)"

The fault is caused because the initial /BP pdfmark is corrupted. This
means that a substream and resource for the 'Picture' has not been
created, leading to us trying to dereference a NULL pointer.

If the resource is NULL< throw an 'undefined' error instead.