Bug 696416: clean up pdfwrite form stack after error.
In pdfwrite, when we start accumulating a form (whether a 'real' form, or a
transparency group), we save a load of 'state' in stack form, so we can then
unroll it when the form execution completes.

If we begin a form, and the form content throws an error, we can reach the end
of the page whilst still having not unrolled the state from the form level back
to the page level, leaving the form stream object as the 'current' stream
rather than the page stream.

This results in the form not being marked as used, and potentially freed,
resulting in a seg fault when we try to flush the current stream (expecting it
to be the page stream, not the form stream).

Before we proceed to complete the page, ensure that we've unrolled any errant
form states.