Bug 697909: Handle PDF T3 glyph streams with unmatched 'q'
We have to handle the case where a glyph description stream from a PDF Type 3
font has an unmatched 'q'. We can track this in the PDF interpreter by checking
the depth of the dictionary stack, and unwinding with 'Q' operations until the
dictionary stack is back to where we started.