Add gs_set_object_type function to gs_memory_t interface.
The subclassing code uses a horribly unsafe hack to reset
the type of a device object. This fails when the device
object is not of the expected type (in this case, when it's
not from the i_alloc manager).

Add a new access function to the API to allow this to happen
safely regardless of manager type.