Remove shared device psi implementation and improve documentation.
The experimental gpdl build now only supports running PostScript and
PDF in a seperate context with it's own device using the normal API.
At this time it only runs files detected as PostScript or PDF in the
default device with no arguments.  The previous setup did not work
properly and if we are to have a shared device it should be
implemented anew anyway.

Also improve the documentation for how languages plug into the
language switching system.