Bug 696377 - Fix a seg fault in subclassing
Bug #696377 "Seg faults found by fuzzing in default_subclass_create_compositor"

With a sufficiently corrupted file it was possible to reach this point
with a mismatch of transparency group begin/end. This could lead to
us having the compositor device being NULL which we would then try to
dereference.

Simply check the compositor device and don't try to dereference it if
its NULL.

This was only reproducible on the commit mentioned in the bug report,
after fixing it the file continued on to seg fault in a completely
different place. However I believe this is due to the fact that the
corrupted file leads to a mismatch of push/pop groups for transparency
and Michael has addressed that with later commits. The file runs to
completion (with *many* errors) on current code.