Bug 696412 - Fix seg fault in pdf_end_transparency_mask
Bug #696412 "Seg faults found by fuzzing in pdf_end_transparency_mask"

The file is (of course) badly corrupted, leading to us receiving a
'end' for a transparency mask when we did not previously receive a
'begin'.

The code in pdf_end_transparency_mask assumed that the prior begin would
have created a soft mask dictionary, but of course if we didn't see the
begin then we don't have one.

Check for this and simply return an error instead of trying to
dereference the NULL pointer.