PDF interpreter - better match the behaviour of Acrobat with invalid /Mask
Bug #697919 "Invalid Mask values cause content to not render"

The sample file has a 1 bit per component image with a Mask, the values
in the mask array are [204 255]. Since this is a 1 bit image, clearly the
only valid values are 0 or 1.

Previously (bug #692852) we have specifically treated /Indexed spaces
with an invalid Mask specially, so that if the initial value is out
of range we ignore the Mask array, otherwise we use the initial value
and set the second to 1.

Also previously we have bug #689717 which is, in fact, an identical
case to this one. In that case we have code which clamps the two values
to 'max - 1' which works for that case. In fact, it works too well for
that case, and produces content which is almost certainly desired by
the author, but which Acrobat does not render.

In this commit we make the DeviceGray case match the /Indexed case, I
rather suspect that for 1 bit images Acrobat *never* uses the /Mask
array values, but I don't have an example in any other colour space. So
for now we'll leave them alone.

This causes the file for Bug 689717 to render with content 'missing',
however it does now match Acrobat. I do not think we can reasonably match
Acrobat's result for this bug without also matching it for the prior bug.