PXL: Optimise 1 pixel wide image handling.
If we are given a 1xH image, we'd rather reflect it to be
an Hx1 image so that it gets passed through the image code
as a single scanline. This allows things like collation of
runs of matching pixels to work far better.

In particular this improves page mode performance for j11.prn.