Bug 697910: handle Type 1 fonts with 'embedded' Type 1 fonts
The test file for this bug has a Type 1 font stream (for an oblique version of
a font) which 'contains' a second (non-oblique) Type 1 font, from which the
'outer' font pulls the CharStrings (the oblique version contains its own
subrs to make the regular font into the oblique font).

A previous fix for a broken Type 1 font broke support for the above (valid)
form of Type 1 font. The broken Type 1 ends up putting various required keys
in the current dictionary, rather than in the font dictionary.

The previous fix pushes a sacrificial "top dictionary" before executing the font
stream. It identifies the broken font by spotting that required keys exist in
a dictionary on the dictionary stack (rather than in the font dictionary on the
operand stack on the call to definefont). When it encounters the broken font
layout, it copies the required keys into that sacrificial top dictionary, and
uses that as the font dictionary.

The Type 1 inside another Type confuses that code, and ends up creating one font
from the "inner" Type 1.

This commit reverses the logic: it checks for required keys missing from the
font dictionary, and if any are missing, it tries to find them on the
dictionary stack.

This is slightly less robust than the original solution, since it assumes that
the font has created a font dictionary, but has defined required keys before
pushing it onto the dictionary stack. A font that fails entirely to create
a font dictionary will fail.

Obviously, a solution for broken fonts that causes (admittedly strange) but
valid fonts to fail is unacceptable. If we find a case of a font that is broken
as described above, we can revisit this fix.