pdfwrite - unescape specially escaped names in pdfmark operations
Bug #697885 "Commented pdf-file converting with 9.21 comments are omitted but not with <= 9.20 versions"

We now preserve (where possible) annotations from input PDF files
into the output file, where previously these were effectively 'folded into'
the page content stream.

In order to do this for Appearances, we need to create a separate Form
XObjet which we reference from the annotation. Doing this requires
some underhandedness in the PDF interpreter, where me make use of
a non-standard 'escape' consisting of tow NULL (0x00) bytes before the
name of the Form XObject, and one NULL after.

This works well, but ended up copying the NULL bytes into the output
PDF file. This is actually legal, but it seriously confused the
Linearisation code when it tried to alter the object numbers in order
to write them in the correct xref table.

Since these NULLs are not required and confuse the code, we now take
extra action to remove them from the string we create working from the
pdfmark data.