device subclassing fix reference counting and stype patching
We patch the memory manager's 'stype' on the original device's memory
when we subclass the device, but had missed resetting that when we
unsubclass the device, restoring the child device. This left the memory
manager stype set incorrectly.

Secondly it transpires that we can alter the reference count of the
parent, so that it differs from the child, and we hadn't accounted for
that. If the reference count of the child (subclassed) device differs
from the parent (subclassing) device at the point where we unsubclass
the device, we need to make sure we retain the correct reference
count. Otherwise we can end up with saved gstates pointing to a device
which has been freed.